Architecture: Re-write core & Render tree :hammer:
1. Previously every `Widget` was creating a `BuildContext` and a `RenderObject`. Then internally to speed DOM rendering, Renderer was also creating a `RenderNode`(kind of a node in Virtual DOM) behind the scenes for every widget. Most of these piece were mutable and to manage them framework was also creating a wrapper object called `WidgetObject` to wrap these objects into a single object. We're finally able to remove all these messy objects from core and what we're left with is two trees, widget tree and a render tree, and there are only two types of objects a widget must have:

    - Widget (immutable node in widget tree that contains configuration for render elements)

    - RenderElement (mutable node in render tree that contains lifecycle methods)

Unlike Element in Flutter, RenderElements are not responsible for mounting or rendering actual DOM nodes. Instead RenderElements can provide a desired description of dom node that they want and how framework renders DOM is totally out of scope for render elements. RenderElements are more like nodes in Virtual DOM from React. A RenderElement can know its parent child relationships but how those child/parents are populated and linked to it are controlled by the framework. This way framework can optimise rendering behind the scenes without breaking APIs.

2. Second change is to Widget's Key property. I've decided to diverge from Flutter's Key model a bit. Unlike flutter, keys now only make sense in the context of the sorrounding array of widgets and there are only two types of keys.

    - Normal key: `Key('requires string value to create key')`

    - Global key: `GlobalKey('requires unique string value to create key')`

Widgets with global key are registered in walker service and a global key can be used to find `RenderElement` associated with a widget using `WalkerService.getRenderElementAssociatedWithGlobalKey`. This is the **only difference** between normal Key and a GlobalKey. e.g A widget with `Key('s')` will be considered matched with a widget that has `GlobalKey('s')`.